CFW Settings toggle MOD
Changed the way CFW settings is toggled so it uses less files and has to
rename less files, for example the rco doesnt need to be renamed, and
now only 1 xml needs to be renamed. also only 1 extra xml needs to be
included instead of 2. It also means it will be easier for any other
modders to edit the network cateogry in the future as the main xmls are
not swapped.